refactor: inline AuthConfig
